Matlab gadgets and example working on unix.
Former-commit-id: 3db9ed5d9cff60e8a0a6cbc5dfb6cca988685ac3

+++ gadgets/matlab/scale.m
@@ -13,7 +13,7 @@ classdef scale < BaseGadget
         function process(g, head, data)
     	    fprintf('Processing line = %d\n', head.idx.kspace_encode_step_1);
             reshdr = head;
-            reshdr.version = 99;
+            reshdr.version = uint16(99);
             resdata = g.factor * data;
             g.putQ(reshdr, resdata);
         end
